bruneimap: Maps and Spatial Data of Brunei

The bruneimap package provides a rich collection of spatial data in sf format for Brunei, including boundaries for districts, mukims, and kampongs, as well as a silhouette map of the country. Additional datasets include locations of masjids, hospitals, clinics, and schools.

See also

Author

Maintainer: Haziq Jamil haziq.jamil@gmail.com (ORCID) [copyright holder]

Other contributors:

Hafeezul Raziq [contributor]

Alvin Bong [contributor]

Aniq Najwa Ariffin [contributor]

Danish Ikhwan Mohamad Zainin [contributor]

Muhammad Rayme Hijazi Jassrie [contributor]

Angela Chan [contributor]

Irdina Batrisyia Norharddyman [contributor]

Nursyafiqah Mohammad Yusuf [contributor]

Ahmad Syafi Mohammad Hilmy [contributor]

Mohammad Syahir Deedat Al Jufri [contributor]