Masjid (mosques) in Brunei

Masjid (mosques) in Brunei

Usage

masjid

Format

A tibble containing the name and lat/long coordinates of masjids in Brunei

Source

Ministry of Religious Affairs Annual Statistics 2021. URL: https://www.mora.gov.bn/SiteCollectionDocuments/Statistik/Perangkaan%20Tahunan%20KHEU%202021.pdf

Senarai Masjid di Negara Brunei Darussalam. URL: <https://www.mora.gov.bn/senaraimasjid/forms/viewall.aspx

Examples

dplyr::glimpse(masjid)
#> Rows: 118
#> Columns: 14
#> $ name           <chr> "Surau Polis Gadong", "Surau Maktab Teknik Sultan Saifu…
#> $ kampong        <fct> Kg. Mata-Mata, Kg. Serusop, Lapangan Terbang Antarabang…
#> $ mukim          <fct> Mukim Gadong B, Mukim Berakas A, Mukim Berakas A, Mukim…
#> $ district       <fct> Brunei Muara, Brunei Muara, Brunei Muara, Temburong, Be…
#> $ longitude      <dbl> 4.907070, 4.946498, 4.944105, 4.707671, 4.586281, 4.585…
#> $ latitude       <dbl> 114.9036, 114.9547, 114.9339, 115.0708, 114.2324, 114.1…
#> $ officiated     <dbl> NA, NA, NA, 1968, 1996, 1964, 1954, 1999, 2017, 1986, 1…
#> $ used           <dbl> NA, 1998, NA, 1967, 1996, 1964, 1954, 1999, 2017, 1986,…
#> $ completed      <dbl> NA, 1999, NA, 1967, 1995, 1963, 1954, 1999, 2017, 1986,…
#> $ capacity       <dbl> 800, 1000, NA, 800, 2900, 2500, 2500, 2000, 3000, 3000,…
#> $ estimated_cost <dbl> NA, 135000, NA, 395000, 14000000, 2600000, 250000, NA, …
#> $ type           <fct> surau, surau, surau, masjid, masjid, masjid, masjid, ma…
#> $ category       <fct> Surau, Surau, Surau, Masjid Utama, Masjid Utama, Masjid…
#> $ notes          <chr> NA, "Commonly said to be in Delima Satu but is within K…